+++ twitter_search/etl/data_collection/search_users.py
@@ -20,9 +20,7 @@ class UserSearcher:
         client: tweepy client
     """
 
-    def __init__(
-        self, location, output_file_users, output_file_tweets, query=None
-    ):
+    def __init__(self, location, output_file_users, output_file_tweets, query=None):
         if query is None:
             self.query = self.query_builder(location)
         else:
@@ -128,9 +126,7 @@ class UserSearcher:
         Returns:
             None
         """
-        self.unique_users_dict = util.remove_duplicate_records(
-            self.total_users_dict
-        )
+        self.unique_users_dict = util.remove_duplicate_records(self.total_users_dict)
 
         util.json_maker(self.output_file_user, self.unique_users_dict)
         print("Total number of users:", len(self.total_users))
@@ -145,9 +141,7 @@ class UserSearcher:
         Returns:
             None
         """
-        self.unique_tweets_dict = util.remove_duplicate_records(
-            self.total_tweets_dict
-        )
+        self.unique_tweets_dict = util.remove_duplicate_records(self.total_tweets_dict)
         util.json_maker(self.output_file_tweets, self.unique_tweets_dict)
         print("Total number of tweets:", len(self.unique_tweets_dict))
