+++ twitter_search/etl/run_search_twitter.py
@@ -191,7 +191,5 @@ class TwitterDataHandler:
         """
         list_reader = ListReader(self.paths["input_file_filter_lists"])
         lists_df = list_reader.create_df()
-        list_filter = ListFilter(
-            lists_df, self.paths["output_file_filter_lists"]
-        )
+        list_filter = ListFilter(lists_df, self.paths["output_file_filter_lists"])
         list_filter.keep_relevant_lists()
