+++ twitter_search/run.py
@@ -2,23 +2,11 @@
 Main function to run the Twitter search and data collection process.
 """
 
-from etl.run_search_twitter import TwitterDataHandler
 from argparse import ArgumentParser
 
-
-def strtobool(val):
-    """Convert a string representation of truth to true (1) or false (0).
-    True values are 'y', 'yes', 't', 'true', 'on', and '1'; false values
-    are 'n', 'no', 'f', 'false', 'off', and '0'.  Raises ValueError if
-    'val' is anything else.
-    """
-    val = val.lower()
-    if val in ("y", "yes", "t", "true", "on", "1"):
-        return 1
-    elif val in ("n", "no", "f", "false", "off", "0"):
-        return 0
-    else:
-        raise ValueError("invalid truth value %r" % (val,))
+# Local imports
+from config_utils.util import strtobool
+from etl.run_search_twitter import TwitterDataHandler
 
 
 def main():
@@ -71,9 +59,7 @@ def main():
 
     print("Building query...")
 
-    twitter_data_handler = TwitterDataHandler(
-        location, account_type, list_needed
-    )
+    twitter_data_handler = TwitterDataHandler(location, account_type, list_needed)
 
     if args.num_iterations:
         print(
